backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
openai
celery
redis
flask-mail

flask-cors
flask-cors
tavily-python
langchain-groq
